add backend message-index & frontend comments

+++ src/main/java/org/java/app/MessageRepo.java
@@ -2,5 +2,5 @@ package org.java.app;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MessageRepo extends JpaRepository<Message, Long> {
+public interface MessageRepo extends JpaRepository<Message, Integer> {
 }
